Don't make a blood decal on low damage

+++ dlls/combat.cpp
@@ -1596,7 +1596,7 @@ void CBaseEntity::TraceBleed(float flDamage, Vector vecDir, TraceResult* ptr, in
 	if (BloodColor() == DONT_BLEED)
 		return;
 
-	if (flDamage == 0)
+	if ((int)flDamage == 0)
 		return;
 
 	if ((bitsDamageType & (DMG_CRUSH | DMG_BULLET | DMG_SLASH | DMG_BLAST | DMG_CLUB | DMG_MORTAR)) == 0)
